Material and working on gridMode 0

+++ src/Modules/Voxelizer/TileVoxel.hpp
@@ -10,7 +10,7 @@ struct TileVoxel
     /* data */
     public:
     
-    int height;
+    double height;
     TileType type;
     Rectangleface top;
     std::vector<Rectangleface> right;
